board-qemu: add private hcall to inform updated interrupt-map phandle
The "interrupt-map" property in each PHB node references the "phandle"
property of the "interrupt-controller" node. QEMU sets this property
to the arbitrary value of 0x1111 in the flattened DT. SLOF then patches
each occurence of this phandle to a SLOF specific value. This value is
then used by the guest OS to setup IRQs for every PCI device plugged to
all PHBs.

In order to support PHB hotplug, this SLOF specific phandle value is
communicated back to QEMU with a new KVMPPC_H_UPDATE_PHANDLE hcall.

Suggested-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Greg Kurz <groug@kaod.org>